+++ src/Interpreter/AbstractCPU.java
@@ -13,9 +13,9 @@ public abstract class AbstractCPU {
 
 	public abstract void run(LinkedList<String[]> instructions) throws Exception;
 	
-	public abstract void setRegister(String register, byte value) throws Exception;
+	public abstract void setRegister(String register, Integer value) throws Exception;
 
-	public abstract byte getRegister(String register);
+	public abstract Integer getRegister(String register);
 	
 	public void enableDebug(boolean isEnabled) {
 		this.debugFlag = isEnabled;
